# docs/api-swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "星云盘 V2 API",
    "description": "星云盘 V2 是一个基于 Go + GORM + Redis + MinIO 的高性能云盘系统。\n\n## 功能特性\n- 用户注册与登录（JWT 认证）\n- 文件上传与下载（支持秒传）\n- 文件夹管理（创建、重命名、移动、删除）\n- 文件去重（基于 SHA256 哈希）\n- 引用计数管理\n\n## 认证方式\n大部分 API 需要在请求头中携带 JWT Token：\n```\nAuthorization: Bearer <access_token>\n```",
    "version": "2.0.0",
    "contact": {
      "name": "星云盘团队",
      "email": "support@xingyunpan.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.xingyunpan.com",
      "description": "生产环境"
    }
  ],
...
  "paths": {
...
    "/api/shares": {
      "post": {
        "tags": ["分享"],
        "summary": "创建分享",
        "description": "创建文件分享链接，支持密码保护和过期时间",
        "operationId": "createShare",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CreateShareRequest"},
              "example": {
                "file_ids": ["file-uuid-1", "file-uuid-2"],
                "expires_in": 7,
                "access_code": "1234"
              }
            }
          }
        },
        "responses": {
          "201": {"description": "创建成功"}
        }
      }
    },
    "/api/shares/me": {
      "get": {
        "tags": ["分享"],
        "summary": "获取我的分享列表",
        "operationId": "getMyShares",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {"description": "获取成功"}
        }
      }
    },
    "/api/shares/{shareId}": {
      "get": {
        "tags": ["分享"],
        "summary": "获取分享信息",
        "operationId": "getShareInfo",
        "parameters": [
          {
            "name": "shareId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {"description": "获取成功"},
          "410": {"description": "分享已过期"}
        }
      },
      "delete": {
        "tags": ["分享"],
        "summary": "删除分享",
        "operationId": "deleteShare",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "shareId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "204": {"description": "删除成功"}
        }
      }
    },
    "/api/shares/{shareId}/verify": {
      "post": {
        "tags": ["分享"],
        "summary": "验证分享密码",
        "operationId": "verifySharePassword",
        "parameters": [
          {
            "name": "shareId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/VerifyPasswordRequest"}
            }
          }
        },
        "responses": {
          "200": {"description": "验证成功"},
          "429": {"description": "超过验证次数限制"}
        }
      }
    },
    "/api/shares/{shareId}/download": {
      "post": {
        "tags": ["分享"],
        "summary": "增加下载计数",
        "operationId": "incrementDownloadCount",
        "parameters": [
          {
            "name": "shareId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "204": {"description": "成功"}
        }
      }
    },
...
    "/api/recycle": {
      "post": {
        "tags": ["回收站"],
        "summary": "移动到回收站",
        "operationId": "moveToRecycle",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/MoveToRecycleRequest"}
            }
          }
        },
        "responses": {
          "204": {"description": "移动成功"}
        }
      },
      "get": {
        "tags": ["回收站"],
        "summary": "获取回收站列表",
        "operationId": "getRecycleList",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "page_size", "in": "query", "schema": {"type": "integer", "default": 20}}
        ],
        "responses": {
          "200": {"description": "获取成功"}
        }
      },
      "delete": {
        "tags": ["回收站"],
        "summary": "永久删除",
        "operationId": "permanentDelete",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/PermanentDeleteRequest"}
            }
          }
        },
        "responses": {
          "204": {"description": "删除成功"}
        }
      }
    },
    "/api/recycle/restore": {
      "post": {
        "tags": ["回收站"],
        "summary": "恢复文件",
        "operationId": "restoreFiles",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/RestoreFilesRequest"}
            }
          }
        },
        "responses": {
          "204": {"description": "恢复成功"}
        }
      }
    },
    "/api/recycle/all": {
      "delete": {
        "tags": ["回收站"],
        "summary": "清空回收站",
        "operationId": "emptyRecycleBin",
        "security": [{"BearerAuth": []}],
        "responses": {
          "204": {"description": "清空成功"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateShareRequest": {
        "type": "object",
        "required": ["file_ids"],
        "properties": {
          "file_ids": {
            "type": "array",
            "items": {"type": "string"},
            "minItems": 1
          },
          "expires_in": {
            "type": "integer",
            "nullable": true
          },
          "access_code": {
            "type": "string",
            "nullable": true,
            "minLength": 4,
            "maxLength": 8
          }
        }
      },
      "VerifyPasswordRequest": {
        "type": "object",
        "required": ["access_code"],
        "properties": {
          "access_code": {"type": "string"}
        }
      },
      "MoveToRecycleRequest": {
        "type": "object",
        "required": ["file_ids"],
        "properties": {
          "file_ids": {
            "type": "array",
            "items": {"type": "integer"},
            "minItems": 1
          }
        }
      },
      "RestoreFilesRequest": {
        "type": "object",
        "required": ["item_ids"],
        "properties": {
          "item_ids": {
            "type": "array",
            "items": {"type": "integer"},
            "minItems": 1
          }
        }
      },
      "PermanentDeleteRequest": {
        "type": "object",
        "required": ["item_ids"],
        "properties": {
          "item_ids": {
            "type": "array",
            "items": {"type": "integer"},
            "minItems": 1
          }
        }
      },
...
    }
  }
}